third_party/blink/web_tests/external/wpt/custom-elements/registries/scoped-custom-element-registry-customelementregistry-attribute-expected.txt

This is a testharness.js-based test.
[FAIL] HTML parser should create a builtin element with null registry if customelementregistry is set
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Cloning a builtin element with null regsitry should create an element with null registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] HTML parser should create a custom element candidate with null registry if customelementregistry is set
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Cloning a custom element candidate with null regsitry should create an element with null registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] HTML parser should create a custom element with null registry if customelementregistry is set
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] Cloning a custom element with null regsitry should create an element with null registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
Harness: the test ran to completion.
